requirements.txt
streamlit==1.34.0
pandas==2.1.4
numpy==1.26.3
plotly==5.18.0
matplotlib==3.8.2
yfinance==0.2.36
scikit-learn==1.4.0
beautifulsoup4==4.12.2
requests==2.31.0
newspaper3k==0.2.8
googletrans==4.0.0-rc1
nltk==3.8.1
textblob==0.17.1
newsapi-python==0.2.7
websocket-client==1.7.0
websockets==12.0
tradingview-ta==3.3.0
investpy==1.0.8
ccxt==4.1.80
passlib==1.7.4
jwt==1.3.1
python-dotenv==1.0.0
streamlit-authenticator==0.2.3
apscheduler==3.10.4